Force non-native SQLite3 in development
Force non-native SQLite3 in development
because there is an issue on CI with the native version and Ruby 3.4
that I was not able to replicate locally.
Probably this will be fixed over time
so the the native version can be used instead,
but it doesn't impact users of the gem either way.